Jornada_Fast_ADOS-Max/Backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
neo4j==5.13.0
